refactor: 기본 home 페이지 정리
