+++ apps/admin/src/app/(developer-guide)/developer-guide-client.tsx
