--- apps/admin/src/app/developer-guide/page.tsx
@@ -1,7 +1,0 @@
-import DeveloperGuideClient from "./developer-guide-client";
-
-export default function DeveloperGuidePage() {
-  return (
-    <DeveloperGuideClient />
-  );
-}
