+++ apps/admin/src/app/page.tsx
@@ -1,23 +1,12 @@
 "use client";
 
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import { login, LoginProvider } from "@/services/auth";
+import DeveloperGuideClient from "./(developer-guide)/developer-guide-client";
 
 export default function Home() {
-  const onLogin = async (provider: LoginProvider) => {
-    const response = await login(provider);
-    console.log("response: ", response);
-  };
+  // const onLogin = async (provider: LoginProvider) => {
+  //   const response = await login(provider);
+  //   console.log("response: ", response);
+  // };
 
-  return (
-    <div className="flex flex-col items-center justify-center h-screen"></div>
-  );
+  return <DeveloperGuideClient />;
 }
