+++ apps/admin/src/components/layout/admin-layout.tsx
@@ -1,12 +1,14 @@
 "use client";
 
-import type React from "react";
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
 import { Menu, X } from "lucide-react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
+import { useState } from "react";
+
+import type React from "react";
+
 import { navItems } from "@/components/layout/sidebar-nav";
+import { Button } from "@/components/ui/button";
 
 interface AdminLayoutProps {
   children: React.ReactNode;
@@ -18,16 +20,24 @@ export function AdminLayout({ children }: AdminLayoutProps) {
 
   return (
     <div className="flex min-h-screen">
-      <div className={`border-r h-screen bg-background transition-all duration-300 ${isCollapsed ? 'w-16' : 'w-64'}`}>
+      <div
+        className={`border-r h-screen bg-background transition-all duration-300 ${isCollapsed ? "w-16" : "w-64"}`}
+      >
         <div className="h-14 border-b flex items-center justify-between px-4">
-          {!isCollapsed && <h2 className="text-lg font-semibold">관리자 시스템</h2>}
+          {!isCollapsed && (
+            <h2 className="text-lg font-semibold">관리자 시스템</h2>
+          )}
           <Button
             variant="ghost"
             size="icon"
             onClick={() => setIsCollapsed(!isCollapsed)}
-            className={isCollapsed ? 'mx-auto' : 'self-end'}
+            className={isCollapsed ? "mx-auto" : "self-end"}
           >
-            {isCollapsed ? <Menu className="h-4 w-4" /> : <X className="h-4 w-4" />}
+            {isCollapsed ? (
+              <Menu className="h-4 w-4" />
+            ) : (
+              <X className="h-4 w-4" />
+            )}
           </Button>
         </div>
         <nav className="space-y-2 p-2">
@@ -39,7 +49,7 @@ export function AdminLayout({ children }: AdminLayoutProps) {
                 href={item.href}
                 key={item.href}
                 aria-label={item.title}
-                className={`flex items-center gap-2 px-3 py-2 rounded-md ${isActive ? 'bg-accent' : 'hover:bg-accent/50'}`}
+                className={`flex items-center gap-2 px-3 py-2 rounded-md ${isActive ? "bg-accent" : "hover:bg-accent/50"}`}
               >
                 <item.icon className="h-5 w-5" />
                 {!isCollapsed && <span>{item.title}</span>}
@@ -52,9 +62,7 @@ export function AdminLayout({ children }: AdminLayoutProps) {
         <header className="h-14 border-b flex items-center justify-between px-6">
           <h1 className="text-lg font-semibold">관리자 페이지</h1>
         </header>
-        <main className="flex-1 p-6 overflow-auto">
-          {children}
-        </main>
+        <main className="flex-1 p-6 overflow-auto">{children}</main>
       </div>
     </div>
   );
